my_blog/requirement.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==2.2
django-allauth==0.44.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-taggit==1.3.0
idna==2.10
importlib-metadata==3.3.0
jsonfield==3.1.0
Markdown==3.3.3
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
Pygments==2.7.3
PyJWT==2.0.0
PyMySQL==0.10.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
